build(protocols): Don't track build output (#1700)
There exists a build cache bug where build restarts from protocols even
when nothing has changed. This is an attempt to fix that bug, though it
is difficult to verify because it does reproduce consistently.

The hypothesis is that because the proto directory is source controlled
Nx assumes that it is source code which causes some sort of issue with
the cache calculation. Since it is build output it does not need to be
source controlled so this removes it from git and ignores it moving
forward.