docs: remove audio from autoplaying video
It was distracting when you navigate to the page
and didn't serve an important purpose since it was only
mouse click sounds, not a voiceover.

The result matches what we do for the other autoplaying video
on http://localhost:8080/guide/getting-started.html,
where the "volume" button is grayed out and can't be clicked.