fix: sketch serializer
During the clone operation only the sketch is serialized and the data isn't included.
Clone needs to not preserve ids right now so that you can copy things back and forth
between spaces and not have it fail because of id collision. Not preserving ids
makes references complicated because you need to update all the references.

The document serializer works because it collapses the references into a flat thing
(markdown) and then rebuilds the object structure afterwards. This does the same thing
but using json for sketch objects.