chore: remove ui packages from paths
When ui packages were included in paths vite was optimizing react-ui in
some cases which lead to there being multiple theme contexts. This
likely points to some other issues with the dependency graph, but this
workaround makes everything work (with the caveat that ui packages need
to be rebuilt by watch to update in composer).

Also removes unused dependencies which were causing emotion to be pulled
into the app bundle.